+++ tests/test_plonecli.py
@@ -10,6 +10,7 @@ from plonecli import cli
 import click
 import os
 import pytest
+import subprocess
 
 
 @pytest.fixture
@@ -50,6 +51,10 @@ def test_command_line_interface():
     version_result = runner.invoke(cli.cli, ['--versions'])
     assert version_result.exit_code == 0
     assert 'Available packages:' in version_result.output
+
+    version_result = runner.invoke(cli.cli, ['build', '-p', 'python3.7'])
+    # assert version_result.exit_code == 0
+    assert 'python3.7' in version_result.output
 
 
 def test_plonecli_test():
@@ -109,3 +114,79 @@ git_init = y
             obj=context.obj,
         )
         assert u'./bin/test --test  src/collective/todo/tests/test_robot.py --package  collective.todo --all' in test_command_result.output  # NOQA: E501
+
+
+def test_plonecli_build_default_py(tmpdir, plonecli_bin):
+    # plonecli_path = os.getcwd()
+    # plonecli_bin = plonecli_path + '/bin/plonecli'
+    target_path = tmpdir.strpath
+    os.chdir(target_path)
+    template = """
+setuptools==40.8.0
+zc.buildout==2.13.1
+"""
+    with open('requirements.txt', 'w') as f:
+        f.write(template)
+
+    template = """[buildout]
+parts =
+"""
+    with open('buildout.cfg', 'w') as f:
+        f.write(template)
+
+    template = """[main]
+version = 5.2-latest
+template = plone_addon
+git_init = y
+"""
+    with open('bobtemplate.cfg', 'w') as f:
+        f.write(template)
+
+    result = subprocess.check_output(
+        [
+            plonecli_bin,
+            'build',
+            '-p',
+            'python3.7',
+        ],
+        cwd=target_path,
+    )
+    assert u'\nRUN: virtualenv . -p python3.7' in result.decode()
+
+
+def test_plonecli_build_py_option(tmpdir, plonecli_bin):
+    # plonecli_path = os.getcwd()
+    # plonecli_bin = plonecli_path + '/bin/plonecli'
+    target_path = tmpdir.strpath
+    os.chdir(target_path)
+    template = """
+setuptools==40.8.0
+zc.buildout==2.13.1
+"""
+    with open('requirements.txt', 'w') as f:
+        f.write(template)
+
+    template = """[buildout]
+parts =
+"""
+    with open('buildout.cfg', 'w') as f:
+        f.write(template)
+
+    template = """[main]
+version = 5.2-latest
+template = plone_addon
+python = python3
+"""
+    with open('bobtemplate.cfg', 'w') as f:
+        f.write(template)
+
+    result = subprocess.check_output(
+        [
+            plonecli_bin,
+            'build',
+            '-p',
+            'python3.7',
+        ],
+        cwd=target_path,
+    )
+    assert u'\nRUN: virtualenv . -p python3' in result.decode()